Added Bellman Ford Shortest Path Algorithms

+++ Algorithms/GraphTheory/Implementations/Edge.java
@@ -3,7 +3,7 @@ package Algorithms.GraphTheory.Implementations;
 public class Edge<E> {
 
     public E from, to;
-    public int weight;
+    public long weight;
 
     public Edge(E from, E to, int weight) {
         this.from = from;
